Change ids to be int64_t and fix type conversion warnings
Job ids can be bigger than a 32 bit int which is supported in this change.
Delay, Priority and TTR have been changed to unsigned ints.

+++ examples/cpp/example1.cc
@@ -12,7 +12,7 @@ int main() {
     assert(client.watch("test"));
     assert(client.ignore("default"));
 
-    int id = client.put("hello");
+    int64_t id = client.put("hello");
     assert(id > 0);
     cout << "put job id: " << id << endl;
